+++ CHANGELOG.md
@@ -15,3 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added datalayer on map to display polygons [#15](https://github.com/Direct4Ag/frontend/issues/15)
 - Add Drought Resistant Seed research component [#19](https://github.com/Direct4Ag/frontend/issues/19)
 - Add Crop Rotation research component [#25](https://github.com/Direct4Ag/frontend/issues/25)
+
+### Fixed
+
+- Change unit from lbs/acre to bu/A in drought resistant seed page [#26](https://github.com/Direct4Ag/frontend/issues/26)